New: Style prop units / derived styles
For now I am hardcoding this (every project will need to copy-paste a few files) because this is a new feature in SDT, and I am not sure how it should be encoded yet. This feature is also relatively code-heavy (as opposed to data-heavy), so I am not sure if it even makes sense to make a generator for it at all.