Fix: Move href, src, and formaction from reflected attributes to regular attributes
The mapping between props and attributes for those keys is not exactly an identity, e.g. if the provided URL is a relative one like , reading the attribute's value will return the same relative URL, but reading the prop will return a computed absolute URL that includes the domain name.