New: Link reflected attrs and props when obtained via ReflectedHtmlAttr
This is needed to give Laminar enough info to remove a prop by removing its reflected html attribute. There is no standard way to unset a built-in JS prop on html elements.

I do not like this approach in SDT very much, but this is what I could muster without refactoring lots of things.